pyproject.toml
[tool.poetry]
name = "fmp-storage"
version = "0.1.0"
description = "Forex Market Predictor storage"
authors = ["bartnyk <jakub@bartnyk.pl>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
annotated-types = "0.7.0"
anyio = "4.6.2.post1"
click = "8.1.7"
dnspython = "2.7.0"
fastapi = "0.115.5"
h11 = "0.14.0"
httptools = "0.6.4"
idna = "3.10"
markdown-it-py = "3.0.0"
mdurl = "0.1.2"
motor = "3.6.0"
pydantic = "2.10.2"
pydantic-settings = "2.6.1"
pydantic-core = "2.27.1"
pygments = "2.18.0"
pymongo = "4.9.2"
python-dotenv = "1.0.1"
pyyaml = "6.0.2"
rich = "13.9.4"
shellingham = "1.5.4"
sniffio = "1.3.1"
starlette = "0.41.3"
typer = "0.14.0"
typing-extensions = "4.12.2"
uvicorn = "0.32.1"
uvloop = "0.21.0"
watchfiles = "1.0.0"
websockets = "14.1"
pytest = "^8.3.4"
ruff = "^0.8.1"
requests = "^2.32.3"
faker = "^33.1.0"
yfinance = "^0.2.50"
isort = "^5.13.2"
pydantic-mongo = "^2.3.0"
selenium-wire = "^5.1.0"
webdriver-manager = "^4.0.2"
blinker = "1.7.0"
setuptools = "^75.7.0"
fmp-core = "^0.1.0"
selenium = "^4.27.1"
logging = "^0.4.9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
